src/test/java/practice/leecode/Tree/IsCompleteTreeTest.java
package practice.leecode.Tree;

import static org.junit.Assert.*;

import org.junit.Test;
import practice.utils.TreeNode;

public class IsCompleteTreeTest {

    @Test
    public void isCompleteTree_1() {
        TreeNode node1 = new TreeNode(1);
        TreeNode node2 = new TreeNode(2);
        TreeNode node3 = new TreeNode(3);
        TreeNode node4 = new TreeNode(4);
        TreeNode node5 = new TreeNode(5);
        TreeNode node6 = new TreeNode(6);

        node1.left = node2;
        node1.right = node3;
        node2.left = node4;
        node2.right = node5;
        node3.left = node6;

        IsCompleteTree IsCompleteTree = new IsCompleteTree();
        assertTrue(IsCompleteTree.isCompleteTree(node1));
    }

    @Test
    public void isCompleteTree_2() {
        TreeNode node1 = new TreeNode(1);
        TreeNode node2 = new TreeNode(2);
        TreeNode node3 = new TreeNode(3);
        TreeNode node4 = new TreeNode(4);
        TreeNode node5 = new TreeNode(5);
        TreeNode node7 = new TreeNode(7);

        node1.left = node2;
        node1.right = node3;
        node2.left = node4;
        node2.right = node5;
        node3.right = node7;

        IsCompleteTree IsCompleteTree = new IsCompleteTree();
        assertFalse(IsCompleteTree.isCompleteTree(node1));
    }

}